Fix bugs in `jsi`
If editing a function and a variable was undeclared in the global scope, `jsi` would attempt to find information about that variable in the global scope, fail to, and decide not to try again. As such, if a variable was later declared in the global scope with the same name, `jsi` would not fetch auto-complete information for it.

Similarly, regardless of cursor position, pressing return/enter would run the script block the user was working on. This has been changed to take cursor position into account.

When `prompt_toolkit` is available, the input system emulated `emacs`. This has been changed to make it emulate `vim`.